refactor(api-types): common schema 위치 및 naming 변경
