+++ packages/api-types/src/common/auth.ts
@@ -1,0 +1,3 @@
+import z from "zod";
+
+export const VerificationCodeSchema = z.string().regex(/^[0-9]{6}$/);
