+++ packages/api-types/src/common/author.ts
@@ -1,0 +1,7 @@
+import z from "zod";
+
+export const AuthorNicknameSchema = z.string().min(2).max(10);
+
+export const AuthorDtoSchema = z.object({
+  nickname: AuthorNicknameSchema,
+});
