--- packages/api-types/src/common/user-profile-dto.ts
@@ -1,8 +1,0 @@
-import z from "zod";
-
-export const UserProfileDtoSchema = z.object({
-  nickname: z.string(),
-  age: z.int(),
-});
-
-export type UserProfileDto = z.output<typeof UserProfileDtoSchema>;
