--- packages/api-types/src/post-comments/common.ts
@@ -1,3 +1,0 @@
-import z from "zod";
-
-export const CommentContentSchema = z.string().min(1).max(200);
