+++ packages/api-types/src/posts/list-posts.ts
@@ -5,7 +5,7 @@
 
 import z from "zod";
 import { responseDtoSchema } from "../helpers";
-import { PostCategorySchema, PostDtoSchema } from "./common";
+import { PostCategorySchema, PostDtoSchema } from "../common/post";
 import { numberSerializerCodec } from "src/common/codecs";
 import { PostErrorCode } from "../common/error-codes";
